App.test.jsx 내부에 있던 컴포넌트를 Restaurants  페이지로 분리
