+++ src/page/Restaurants.jsx
@@ -1,0 +1,26 @@
+import { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
+
+import RegionsContainer from '../RegionsContainer';
+import CategoriesContainer from '../CategoriesContainer';
+import RestaurantsContainer from '../RestaurantsContainer';
+
+import {
+  loadInitialData,
+} from '../actions';
+
+export default function Restaurants() {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(loadInitialData());
+  });
+
+  return (
+    <div>
+      <RegionsContainer />
+      <CategoriesContainer />
+      <RestaurantsContainer />
+    </div>
+  );
+}
